Improve compatibility with ScienceCostTweakerM
The mod ScienceCostTweakerM [1] changes the ingredients for the science
packs.  When constructing the recipe for our science fluids we copy
those ingredients.  Here order matters:

- first ScienceCostTweakerM must change the science pack ingredients
- then we must derive the fluid recipe and replace the science pack
  recipe with our own.

Otherwise we end up with the ingredients of the old pack recipe for our
fluids and ScienceCostTweakerM overrides our science pack recipe with
their own.

Add ScienceCostTweakerM as optional dependency, to force the correct
ordering.

[1] https://mods.factorio.com/mod/ScienceCostTweakerM